+++ riddle-renderer-wgpu/src/wgpu_ext/buffered_renderer.rs
@@ -1,15 +1,15 @@
-use crate::{math::*, *};
+use crate::wgpu_ext::*;
 
 use wgpu::util::DeviceExt;
 
 #[derive(Clone)]
 pub(crate) struct BufferedRenderArgs {
-    pub(crate) texture: TextureHandle,
-    pub(crate) shader: ShaderHandle,
+    pub(crate) texture: WGPUTextureHandle,
+    pub(crate) shader: WGPUShaderHandle,
 }
 
 impl BufferedRenderArgs {
-    fn new(desc: &RenderableDesc) -> Self {
+    fn new(desc: &WGPURenderableDesc) -> Self {
         Self {
             texture: desc.texture.clone(),
             shader: desc.shader.clone(),
@@ -19,14 +19,18 @@ impl BufferedRenderArgs {
 
 impl PartialEq for BufferedRenderArgs {
     fn eq(&self, other: &Self) -> bool {
-        TextureHandle::eq(&self.texture, &other.texture)
-            && ShaderHandle::eq(&self.shader, &other.shader)
+        WGPUTextureHandle::eq(&self.texture, &other.texture)
+            && WGPUShaderHandle::eq(&self.shader, &other.shader)
     }
 }
 
 impl Eq for BufferedRenderArgs {}
 
-pub(crate) struct BufferedRenderer<'a, R: RenderTargetDesc<'a>> {
+pub(crate) struct BufferedRenderer<'a, Device, R>
+where
+    Device: WGPUDevice,
+    R: WGPURenderTargetDesc<'a, Device>,
+{
     target_desc: R,
     current_args: Option<BufferedRenderArgs>,
 
@@ -34,13 +38,17 @@ pub(crate) struct BufferedRenderer<'a, R: RenderTargetDesc<'a>> {
     indices: Vec<u16>,
 
     pending_clear_color: Option<[f32; 4]>,
-    view_matrix_stack: Vec<mint::ColumnMatrix4<f32>>,
+    view_matrix: mint::ColumnMatrix4<f32>,
 
     encoder: wgpu::CommandEncoder,
-    _a_marker: std::marker::PhantomData<&'a ()>,
+    _a_marker: std::marker::PhantomData<&'a Device>,
 }
 
-impl<'a, R: RenderTargetDesc<'a>> BufferedRenderer<'a, R> {
+impl<'a, Device, R> BufferedRenderer<'a, Device, R>
+where
+    Device: WGPUDevice,
+    R: WGPURenderTargetDesc<'a, Device>,
+{
     pub fn new(target_desc: R, encoder: wgpu::CommandEncoder) -> Result<Self> {
         target_desc.begin_render()?;
         let identity: mint::ColumnMatrix4<f32> = glam::Mat4::identity().into();
@@ -50,7 +58,7 @@ impl<'a, R: RenderTargetDesc<'a>> BufferedRenderer<'a, R> {
             verts: vec![],
             indices: vec![],
             pending_clear_color: None,
-            view_matrix_stack: vec![identity],
+            view_matrix: identity,
             encoder,
             _a_marker: std::marker::PhantomData::default(),
         })
@@ -91,62 +99,65 @@ impl<'a, R: RenderTargetDesc<'a>> BufferedRenderer<'a, R> {
     }
 
     fn do_flush(&mut self, args: &BufferedRenderArgs) -> Result<()> {
-        let device = self.target_desc.wgpu_device().device();
+        let renderer = self.target_desc.renderer().clone_handle();
 
-        let vertex_buf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: None,
-            contents: bytemuck::cast_slice(&self.verts),
-            usage: wgpu::BufferUsage::VERTEX,
-        });
+        let verts = std::mem::take(&mut self.verts);
+        let indices = std::mem::take(&mut self.indices);
+        let pending_clear = self.pending_clear_color.take();
 
-        let index_buf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: None,
-            contents: bytemuck::cast_slice(&self.indices),
-            usage: wgpu::BufferUsage::INDEX,
-        });
+        renderer.wgpu_device().with_device_info(|info| {
+            let vertex_buf = info
+                .device
+                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                    label: None,
+                    contents: bytemuck::cast_slice(&verts),
+                    usage: wgpu::BufferUsage::VERTEX,
+                });
 
-        let view_matrix = self
-            .view_matrix_stack
-            .last()
-            .map(|m| m.clone())
-            .unwrap_or_else(|| glam::Mat4::identity().into());
+            let index_buf = info
+                .device
+                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
+                    label: None,
+                    contents: bytemuck::cast_slice(&indices),
+                    usage: wgpu::BufferUsage::INDEX,
+                });
 
-        let bind_group = args.shader.bind_params(
-            device,
-            self.target_desc.dimensions(),
-            view_matrix,
-            &args.texture,
-        );
+            let bind_group = args.shader.bind_params(
+                info.device,
+                self.target_desc.dimensions(),
+                self.view_matrix,
+                &args.texture,
+            );
 
-        let load_op = match self.pending_clear_color {
-            Some(c) => wgpu::LoadOp::Clear(wgpu::Color {
-                r: c[0] as f64,
-                g: c[1] as f64,
-                b: c[2] as f64,
-                a: c[3] as f64,
-            }),
-            None => wgpu::LoadOp::Load,
-        };
-        self.pending_clear_color = None;
+            let load_op = match pending_clear {
+                Some(c) => wgpu::LoadOp::Clear(wgpu::Color {
+                    r: c[0] as f64,
+                    g: c[1] as f64,
+                    b: c[2] as f64,
+                    a: c[3] as f64,
+                }),
+                None => wgpu::LoadOp::Load,
+            };
 
-        // Scope render pass so that the render is encoded before the buffers are cleared.
-        let encoder = &mut self.encoder;
-        let indices_len = self.indices.len() as u32;
-        self.target_desc.with_view(|view| {
-            let mut rpass = args.shader.begin_render_pass(view, encoder, load_op);
+            // Scope render pass so that the render is encoded before the buffers are cleared.
+            let encoder = &mut self.encoder;
+            let indices_len = indices.len() as u32;
+            self.target_desc.with_view(|view| {
+                let mut rpass = args.shader.begin_render_pass(view, encoder, load_op);
 
-            rpass.set_bind_group(0, &bind_group, &[]);
+                rpass.set_bind_group(0, &bind_group, &[]);
 
-            rpass.set_vertex_buffer(0, vertex_buf.slice(..));
-            rpass.set_index_buffer(index_buf.slice(..));
-            rpass.draw_indexed(0..indices_len, 0, 0..1);
+                rpass.set_vertex_buffer(0, vertex_buf.slice(..));
+                rpass.set_index_buffer(index_buf.slice(..));
+                rpass.draw_indexed(0..indices_len, 0, 0..1);
+                Ok(())
+            })?;
+
+            self.verts.clear();
+            self.indices.clear();
+            self.current_args = None;
             Ok(())
-        })?;
-
-        self.verts.clear();
-        self.indices.clear();
-        self.current_args = None;
-        Ok(())
+        })
     }
 
     pub fn buffered_render(
@@ -172,28 +183,17 @@ impl<'a, R: RenderTargetDesc<'a>> BufferedRenderer<'a, R> {
     }
 }
 
-impl<'a, R: RenderTargetDesc<'a>> RenderContext for BufferedRenderer<'a, R> {
+impl<'a, Device, R> RenderContext for BufferedRenderer<'a, Device, R>
+where
+    Device: WGPUDevice,
+    R: WGPURenderTargetDesc<'a, Device>,
+{
     fn set_transform(&mut self, transform: mint::ColumnMatrix4<f32>) -> Result<()> {
         self.flush()?;
-        match self.view_matrix_stack.last_mut() {
-            Some(last) => *last = transform,
-            _ => (),
-        }
+        self.view_matrix = transform;
         Ok(())
     }
-    /*
-        fn push_transform(&mut self, transform: mint::ColumnMatrix4<f32>) -> Result<()> {
-            self.flush()?;
-            self.view_matrix_stack.push(transform);
-            Ok(())
-        }
 
-        fn pop_transform(&mut self) -> Result<()> {
-            self.flush()?;
-            self.view_matrix_stack.pop();
-            Ok(())
-        }
-    */
     /// Set the clear color and mark the frame buffer for clearing. The actual clear operation
     /// will be performed when the next batched render happens, or when `present` is called,
     /// whichever comes first.
@@ -236,14 +236,20 @@ impl<'a, R: RenderTargetDesc<'a>> RenderContext for BufferedRenderer<'a, R> {
         }
 
         let cmd = self.encoder.finish();
-        self.target_desc.wgpu_device().queue().submit(Some(cmd));
+        self.target_desc
+            .renderer()
+            .wgpu_device()
+            .with_device_info(|info| {
+                info.queue.submit(Some(cmd));
+                Ok(())
+            })?;
 
         self.target_desc.end_render();
 
         Ok(())
     }
 
-    fn render_internal<S: Renderable>(&mut self, renderable: &S) -> Result<()> {
+    fn render_internal<S: WGPURenderable>(&mut self, renderable: &S) -> Result<()> {
         renderable.with_renderable(|r| {
             self.buffered_render(&BufferedRenderArgs::new(r), r.verts, r.indices)?;
             Ok(())