+++ riddle-renderer-wgpu/src/wgpu_ext/sprite.rs
@@ -1,6 +1,4 @@
-use crate::{ext::*, math::*, *};
-
-use riddle_common::Color;
+use crate::wgpu_ext::*;
 
 /////////////////////////////////////////////////////////////////////////////
 // struct Sprite
@@ -9,10 +7,10 @@ use riddle_common::Color;
 /// A renderable region of a texture.
 ///
 /// Multiple sprites can share a single texture. Sprites can either be built using
-/// [`SpriteBuilder`], or [`SpriteAtlasBuilder`].
+/// [`crate::SpriteBuilder`], or [`SpriteAtlasBuilder`].
 ///
-/// Use [`SpriteRenderCommand`] for access to all supported paramters when rendering
-/// sprites, or use [`Sprite::render_at`] to specify only a location and use default
+/// Use [`crate::SpriteRenderCommand`] for access to all supported paramters when rendering
+/// sprites, or use [`WGPUSprite::render_at`] to specify only a location and use default
 /// arguments for everything else.
 ///
 /// Sprites store a reference to the [`Renderer`] which built it, which will keep
@@ -28,7 +26,7 @@ use riddle_common::Color;
 /// let renderer = Renderer::new_from_window(&window)?;
 ///
 /// // Load an image and create a sprite from it
-/// let png_bytes = include_bytes!("../../example_assets/image.png");
+/// let png_bytes = include_bytes!("../../../example_assets/image.png");
 /// let img = Image::load(&png_bytes[..], ImageFormat::Png)?;
 /// let sprite = SpriteBuilder::new(img).build(&renderer)?;
 ///
@@ -39,33 +37,38 @@ use riddle_common::Color;
 /// render_ctx.present()?;
 /// # Ok(()) }
 /// ```
-pub struct Sprite {
-    renderer: RendererHandle,
-    texture: TextureHandle,
+pub struct WGPUSprite<Device: WGPUDevice> {
+    renderer: WGPURendererHandle<Device>,
+    texture: WGPUTextureHandle,
     source_rect: Rect<f32>,
 }
 
-impl Sprite {
+impl<Device: WGPUDevice> WGPUSprite<Device> {
     /// Construct a new sprite from an image. The image contents are copied to a texture
     /// in RGBA8 format. The entire image will be used
     pub(crate) fn new_from_image(
-        renderer: &Renderer,
+        renderer: &WGPURenderer<Device>,
         img: image::Image,
         mag_filter: FilterMode,
         min_filter: FilterMode,
-    ) -> Result<Sprite> {
-        let texture = Texture::from_image(
-            renderer.wgpu_device().device(),
-            renderer.wgpu_device().queue(),
-            img,
-            mag_filter,
-            min_filter,
-            TextureType::Plain,
-        )?;
+    ) -> Result<Self> {
+        let texture = renderer.wgpu_device().with_device_info(|info| {
+            Ok(WGPUTexture::from_image(
+                info.device,
+                info.queue,
+                img,
+                mag_filter,
+                min_filter,
+                TextureType::Plain,
+            )?)
+        })?;
         Self::from_texture(renderer, &texture)
     }
 
-    pub(super) fn from_texture(renderer: &Renderer, texture: &Texture) -> Result<Sprite> {
+    pub(crate) fn from_texture(
+        renderer: &WGPURenderer<Device>,
+        texture: &WGPUTexture,
+    ) -> Result<Self> {
         let dimensions = texture.dimensions.convert();
         Self::from_texture_with_bounds(
             renderer,
@@ -77,12 +80,12 @@ impl Sprite {
         )
     }
 
-    pub(super) fn from_texture_with_bounds(
-        renderer: &Renderer,
-        texture: &Texture,
+    pub(crate) fn from_texture_with_bounds(
+        renderer: &WGPURenderer<Device>,
+        texture: &WGPUTexture,
         source_rect: Rect<f32>,
-    ) -> Result<Sprite> {
-        Ok(Sprite {
+    ) -> Result<Self> {
+        Ok(WGPUSprite {
             renderer: renderer.clone_handle(),
             texture: texture.clone_handle(),
             source_rect,
@@ -119,11 +122,11 @@ impl Sprite {
     /// assert_eq!(vec2(25.0, 25.0), subsprite.dimensions());
     /// # Ok(()) }
     /// ```
-    pub fn subsprite(&self, source_rect: &Rect<f32>) -> Sprite {
+    pub fn subsprite(&self, source_rect: &Rect<f32>) -> Self {
         let mut translated_source = source_rect.clone();
         translated_source.location = translated_source.location + self.source_rect.location;
 
-        Sprite {
+        WGPUSprite {
             renderer: self.renderer.clone(),
             texture: self.texture.clone(),
             source_rect: self
@@ -186,7 +189,7 @@ impl Sprite {
 
         let index_data: &[u16] = &[1, 2, 0, 2, 0, 3];
 
-        let renderable = RenderableDesc {
+        let renderable = WGPURenderableDesc {
             texture: self.texture.clone(),
             shader: self.renderer.standard_res().default_shader.clone(),
             verts: &vertex_data[..],
@@ -235,163 +238,5 @@ impl Sprite {
     /// ```
     pub fn dimensions(&self) -> Vector2<f32> {
         self.source_rect.dimensions
-    }
-}
-
-/////////////////////////////////////////////////////////////////////////////
-// struct SpriteRenderCommand
-/////////////////////////////////////////////////////////////////////////////
-
-/// Builder for a [`Sprite`] render call
-///
-/// Defaults:
-///
-/// * **Pivot**: `(0,0)`
-/// * **Scale**: `(1,1)`
-/// * **Angle**: `0`
-/// * **Diffuse Color**: `Color::WHITE`
-///
-/// The location refers to the location of the pivot of the sprite. The pivot
-/// of the sprite is relative to the top left of the sprite.
-///
-/// # Example
-///
-/// ```no_run
-/// # use riddle::{common::Color, image::*, platform::*, renderer::*, math::*, *};
-/// # fn main() -> Result<(), RiddleError> {
-/// # let rdl =  RiddleLib::new()?;
-/// # let window = WindowBuilder::new().build(rdl.context())?;
-/// # let renderer = Renderer::new_from_window(&window)?;
-/// # let img = Image::new(100, 100);
-/// let sprite = SpriteBuilder::new(img).build(&renderer)?;
-///
-/// let mut render_ctx = renderer.begin_render()?;
-/// render_ctx.clear(Color::WHITE)?;
-///
-/// // Render the sprite
-/// SpriteRenderCommand::new(vec2(0.0, 0.0))
-///     .with_scale(vec2(1.0, 2.0))
-///     .with_color(Color::RED)
-///     .render(&mut render_ctx, &sprite)?;
-///
-/// render_ctx.present()?;
-/// # Ok(()) }
-/// ```
-#[derive(Clone, Debug)]
-pub struct SpriteRenderCommand {
-    location: Vector2<f32>,
-    pivot: Vector2<f32>,
-    scale: Vector2<f32>,
-    angle: f32,
-    diffuse_color: Color<f32>,
-}
-
-impl SpriteRenderCommand {
-    /// New render command with default args, at the specified location
-    pub fn new<T: Into<Vector2<f32>>>(location: T) -> Self {
-        let mut args = Self::default();
-        args.at(location);
-        args
-    }
-
-    /// Set the location of the sprite, specifying where the pivot should
-    /// be placed.
-    #[inline]
-    pub fn at<T: Into<Vector2<f32>>>(&mut self, location: T) -> &mut Self {
-        self.location = location.into();
-        self
-    }
-
-    /// Set the pivot of the sprite, relative to the top left of the sprite
-    #[inline]
-    pub fn with_pivot<T: Into<Vector2<f32>>>(&mut self, pivot: T) -> &mut Self {
-        self.pivot = pivot.into();
-        self
-    }
-
-    /// Set the scale at which the sprite will be rendered
-    pub fn with_scale<T: Into<Vector2<f32>>>(&mut self, scale: T) -> &mut Self {
-        self.scale = scale.into();
-        self
-    }
-
-    /// Set the angle at which the sprite will be rendered, in radians.
-    pub fn with_angle(&mut self, angle: f32) -> &mut Self {
-        self.angle = angle;
-        self
-    }
-
-    /// Set the diffuse color of the sprite, which will be multiplied by the sprite
-    /// colors.
-    pub fn with_color(&mut self, color: Color<f32>) -> &mut Self {
-        self.diffuse_color = color;
-        self
-    }
-
-    /// Invoke the render command, for the given sprite, in the specified context
-    pub fn render(&self, render_ctx: &mut impl RenderContext, sprite: &Sprite) -> Result<()> {
-        sprite.render(render_ctx, self)
-    }
-}
-
-impl Default for SpriteRenderCommand {
-    fn default() -> Self {
-        SpriteRenderCommand {
-            location: [0.0, 0.0].into(),
-            pivot: [0.0, 0.0].into(),
-            angle: 0.0,
-            scale: [1.0, 1.0].into(),
-            diffuse_color: Color::WHITE,
-        }
-    }
-}
-
-/////////////////////////////////////////////////////////////////////////////
-// struct SpriteBuilder
-/////////////////////////////////////////////////////////////////////////////
-
-/// Builder to construct new [`Sprite`]s from `riddle_image::Image`s.
-///
-/// # Example
-///
-/// ```no_run
-/// # use riddle::{common::Color, image::*, platform::*, renderer::*, math::*, *};
-/// # fn main() -> Result<(), RiddleError> {
-/// # let rdl =  RiddleLib::new()?;
-/// # let window = WindowBuilder::new().build(rdl.context())?;
-/// let renderer = Renderer::new_from_window(&window)?;
-///
-/// let img = Image::new(100, 100);
-/// let sprite = SpriteBuilder::new(img)
-///     .with_filter_modes(FilterMode::Linear, FilterMode::Linear)
-///     .build(&renderer)?;
-/// # Ok(()) }
-/// ```
-pub struct SpriteBuilder {
-    img: image::Image,
-    mag_filter: FilterMode,
-    min_filter: FilterMode,
-}
-
-impl SpriteBuilder {
-    /// Create a new builder for the given image
-    pub fn new(img: image::Image) -> Self {
-        Self {
-            img,
-            mag_filter: Default::default(),
-            min_filter: Default::default(),
-        }
-    }
-
-    /// Specify the min and mag filters used when rendering the sprite
-    pub fn with_filter_modes(mut self, mag_filter: FilterMode, min_filter: FilterMode) -> Self {
-        self.mag_filter = mag_filter;
-        self.min_filter = min_filter;
-        self
-    }
-
-    /// Build the sprite for the given renderer
-    pub fn build(self, renderer: &Renderer) -> Result<Sprite> {
-        Sprite::new_from_image(renderer, self.img, self.mag_filter, self.min_filter)
     }
 }